/**  * This code written by Claude Sonnet 4 (claude-3-5-sonnet-20241022)  * Generated via Cursor IDE (cursor.sh) with AI assistance  * Model: Anthropic Claude 3.5 Sonnet  * Generation timestamp: 2024-12-19  * Context: Fix PonyORM entity definitions and restructure email templates  *  * Technical details:  * - LLM: Claude 3.5 Sonnet (2024-10-22)  * - IDE: Cursor (cursor.sh)  * - Generation method: AI-assisted pair programming  * - Code style: Python with full mypy typing and numpy-style docstrings  * - Dependencies: jinja2, pony, toml, click, rich, pathlib  *  * Fixes include:  * - Fixed PonyORM entity definitions (db.Entity instead of Entity)  * - Removed unsupported index() calls from PonyORM models  * - Restructured email templates into separate .html and .txt files  * - Updated template manager to load from files instead of embedded strings  * - Added custom strftime filter for Jinja2 templates  * - Fixed Click type from Decimal to FLOAT  * - Fixed import paths for price watch models  * - Enhanced email config to parse existing email.txt format  * - Added DreamHost SMTP configuration support  */

+++ meijer/cli/price_watch_commands.py
@@ -64,7 +64,7 @@ def price_watch_group():
 )
 @click.option(
     "--desired-price",
-    type=click.Decimal,
+    type=click.FLOAT,
     help="Target price for alerts"
 )
 @click.option(
